pyproject.toml
[tool.poetry]
name = "GH Systems"
version = "0.1.0"
description = "GH Systems REST API"
authors = ["Hugo Silva <hugofer93@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.12"
python-dotenv = "~1.0"
setuptools = "~75.1"
psycopg2 = "~2.9"
Django = "~4.2"
djangorestframework = "~3.15"
drf-spectacular = "~0.27"

[tool.poetry.group.dev.dependencies]
django-silk = "~5.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
